test: isolate followup draft helper
