+++ packages/app/src/components/prompt-input.tsx
@@ -39,7 +39,8 @@ import { pickAttachments } from "./prompt-input/pick-attachments"
 import { ACCEPTED_FILE_TYPES } from "./prompt-input/files"
 import { promptLength } from "./prompt-input/history"
 import type { PromptStore } from "./prompt-input/store-types"
-import { createPromptSubmit, type FollowupDraft } from "./prompt-input/submit"
+import type { FollowupDraft } from "./prompt-input/followup-draft"
+import { createPromptSubmit } from "./prompt-input/submit"
 import { PromptPopover } from "./prompt-input/slash-popover"
 import { PromptContextItems } from "./prompt-input/context-items"
 import { PromptImageAttachments } from "./prompt-input/image-attachments"
